% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{wulab_colors}
\alias{wulab_colors}
\alias{show_color_qualitative}
\alias{show_color_sequential}
\alias{show_color_diverging}
\title{Wu Lab Color Palettes and Visualization}
\usage{
show_color_qualitative()

show_color_sequential(n = 9)

show_color_diverging(n = 9)
}
\arguments{
\item{n}{Integer. Number of colors. Odd numbers recommended.}
}
\description{
A suite of functions to visualize and select lab-standard color palettes.
These palettes are designed for scientific publications,
ensuring consistency in experimental group mapping and data gradients.

\code{show_color_qualitative}: View 12 paired qualitative
colors and 3 grey colors.

\code{show_color_sequential}: View sequential colors for continuous data.

\code{show_color_diverging}: View diverging colors for centered data.
}
\details{
The functions in this family provide a visual reference in the RStudio
Plots pane while simultaneously printing the exact HEX codes to the
console for easy copy-pasting into \code{ggplot2} manual scales.
}
\section{Palettes included}{

\itemize{
  \item \bold{Qualitative}: 12 paired colors inspired by traditional Chinese
  aesthetics, ideal for categorical groups (e.g., WT vs. Mutant). Includes
  3 specialized grey colors for background or missing data.
  \item \bold{Sequential}: A multi-hue gradient from Creamy Avocado (#d9ed92)
  to Moroccan Blue (#184e77) via a Teal midpoint (#52b69a).
  Define the number of colors with the parameter n (Default n = 9).
  \item \bold{Diverging}: A transition from Copper Hopper (#bb3e03) to
  Deep Lagoon (#005f73) with a Shalimar (#ffffbf) midpoint.
  Define the number of colors with the parameter n (Default n = 9).
  Odd numbers work best for a central midpoint.
}
}

